pyproject.toml
[project]
name = "SFX"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-restful>=0.3.10",
    "markdown>=3.9",
    "python-dotenv>=1.1.1",
]
